@@ -1,28 +1,23 @@
-var blacklist = require('blacklist');
-var classnames = require('classnames');
-var React = require('react');
+import classnames from 'classnames';
+import React, { PropTypes } from 'react';
 
-module.exports = React.createClass({
-	displayName: 'Form',
-	propTypes: {
-		children: React.PropTypes.node.isRequired,
-		className: React.PropTypes.string,
-		component: React.PropTypes.oneOfType([
-			React.PropTypes.element,
-			React.PropTypes.string
-		]),
-		type: React.PropTypes.oneOf(['basic', 'horizontal', 'inline'])
-	},
-	getDefaultProps () {
-		return {
-			component: 'form',
-			type: 'basic'
-		};
-	},
-	render () {
-		var props = blacklist(this.props, 'children', 'type');
-		props.className = classnames('Form', ('Form--' + this.props.type), this.props.className);
+function Form ({ className, component, type, ...props }) {
+	const Component = component;
+	props.className = classnames('Form', ('Form--' + type), className);
 
-		return React.createElement(this.props.component, props, this.props.children);
-	},
-});
+	return <Component {...props} />;
+};
+
+Form.propTypes = {
+	component: PropTypes.oneOfType([
+		React.PropTypes.string,
+		React.PropTypes.func,
+	]),
+	type: PropTypes.oneOf(['basic', 'horizontal', 'inline']),
+};
+Form.defaultProps = {
+	component: 'form',
+	type: 'basic',
+};
+
+module.exports = Form;
