@@ -1,5 +1,6 @@
 var React = require('react');
 var classNames = require('classnames');
+var blacklist = require('blacklist');
 
 module.exports = React.createClass({
 	displayName: 'InputGroupSection',
@@ -12,9 +13,10 @@ module.exports = React.createClass({
 		var className = classNames('InputGroup_section', {
 			'InputGroup_section--grow': this.props.grow
 		}, this.props.className);
+		var props = blacklist(this.props, 'grow');
 
 		return (
-			<div {...this.props} className={className} />
+			<div {...props} className={className} />
 		);
 	}
 });
