@@ -1,5 +1,6 @@
 var classnames = require('classnames');
 var React = require('react');
+var blacklist = require('blacklist');
 
 module.exports = React.createClass({
 	displayName: 'ModalHeader',
@@ -20,8 +21,10 @@ module.exports = React.createClass({
 		var className = classnames('Modal__header', this.props.className);
 		var close = this.props.showCloseButton ? <button type="button" onClick={this.handleClose} className="Modal__header__close" /> : null;
 		var text = this.props.text ? <h4 className="Modal__header__text">{this.props.text}</h4> : null;
+		var props = blacklist(this.props, 'children', 'onClose', 'showCloseButton', 'text');
+
 		return (
-			<div {...this.props} className={className}>
+			<div {...props} className={className}>
 				{close}
 				{text}
 				{this.props.children}
