@@ -69,7 +69,7 @@ module.exports = React.createClass({
 		var self = this;
 
 		// props
-		var props = blacklist(this.props, 'alwaysValidate', 'label', 'onChange', 'options', 'required', 'requiredMessage', 'value');
+		var props = blacklist(this.props, 'alwaysValidate', 'label', 'onChange', 'options', 'required', 'requiredMessage', 'value', 'inline');
 
 		// classes
 		var componentClass = classNames('FormField', {
